Remove the optimizer and the asm-all and asm-debug-all jars.
The jar size is no longer an issue for the vast majority of the ASM users, unlike the lack of debug info in the default jars. This changes keeps the debug info in the default jars by removing the optimizer. The asm-debug-all jars is then no longer necessary, and is thus removed.

Similarly, maintaining compatibility with the JDK 1.3 is no longer necessary (it is unsupported since a long time). Instead, this change uses the JDK 1.5 as the new baseline. This is now striclty enforced (the build will fail if a more recent API is used internally), for all jars (previously different jars were using different baseline APIs).