Fix the handling of inner classes with EXPAND_ASM_INSNS.
Resetting the inner class data structures so that they are rebuilt when revisiting the code with a ClassReader does not work, due to the code in visitInnerClass to avoid duplicate entries in this table (based on state stored in the constant pool entries, which are not reset). In fact, thanks to this code to avoid duplicate entries, we can skip resetting the inner class data structures here, and be assured that they will not change when revisiting the class with a ClassReader.