Fix the sigtest task.
The signature test was not working without the rt.jar on the classpath (no error was reported for incompatible API changes). Adding it solves this issue, and also removes the warnings about the @Deprecated annotations. I also had to manually fix the asm-xml signature files (for some unknown reason the inner ASMContentHandler$Rule class was recorded as public, whereas it has always been protected).